+++ redbox-core/tests/retriever/data.py
@@ -9,7 +9,9 @@ ALL_CHUNKS_RETRIEVER_CASES = [
     test_case
     for generator in [
         generate_test_cases(
-            query=ChainInput(question="Irrelevant Question", file_uuids=[uuid4()], user_uuid=uuid4(), chat_history=[]),
+            query=ChainInput(
+                question="Irrelevant Question", file_uuids=[uuid4()], user_uuid=uuid4(), chat_history=[], ai_settings={}
+            ),
             test_data=[TestData(8, 8000)],
             test_id="Successful Path",
         )
@@ -21,7 +23,9 @@ PARAMETERISED_RETRIEVER_CASES = [
     test_case
     for generator in [
         generate_test_cases(
-            query=ChainInput(question="Irrelevant Question", file_uuids=[uuid4()], user_uuid=uuid4(), chat_history=[]),
+            query=ChainInput(
+                question="Irrelevant Question", file_uuids=[uuid4()], user_uuid=uuid4(), chat_history=[], ai_settings={}
+            ),
             test_data=[TestData(8, 8000, ChunkResolution.normal)],
             test_id="Successful Path",
         )
